welp...
The good news is, I have figured out about 432 different ways to NOT do things. So, in a sense....Progress!

 I need to figure out how to properly create rl inside of the GSM, then send it as a parameter to the gamestates. However, If I am doing that, I think I technically wouldn't need to use the DLF to send rl to each location, just the gsm. This leads me to wonder, why I need to import GSM or RL stuff at all within the locations themselves. So that's why I messed around with the monster-room-location. I tried this on a larger scale and broke a bunch of stuff so I commented out a bunch of experimental stuff from other locations.

There is an issue in the GSM where one of my `this`es went rogue when I implemented the changes to the menu-state-constructor.

Oh and the menu is still broke.